backend/requirements.txt
langgraph
langgraph-supervisor
langchain
langchain-tavily
langchain-google-genai
fastmcp
fastapi
uvicorn[standard]
langchain-mcp-adapters
python-dotenv
mcp
pydantic
python-multipart
langgraph-checkpoint-postgres
psycopg2-binary
sqlalchemy